+++ src/extension/blockchainIdentifier.ts
@@ -5,7 +5,7 @@ import * as vscode from "vscode";
 import BlockchainType from "./blockchainType";
 import IoHelpers from "./util/ioHelpers";
 import JSONC from "./util/JSONC";
-import Log from "../shared/log";
+import Log from "./util/log";
 import posixPath from "./util/posixPath";
 
 const LOG_PREFIX = "BlockchainIdentifier";